+++ src/components/dashboard-widgets/OverviewWidget.vue
@@ -1,69 +1,27 @@
 <template>
   <div class="h-full overflow-scroll">
-    <a
-      :href="`${generateUrl('apps/money')}/#/book/${book.id}`"
-      class="grid h-12 grid-cols-2 items-center rounded-full px-2 hover:bg-background-hover"
+    <BookListItem
       v-for="book of bookStore.books.value"
+      :book="book"
+      :href="`${generateUrl('apps/money')}/#/book/${book.id}`"
       :key="book.id"
-    >
-      <div class="truncate font-semibold">
-        {{ book.name || `(${t('money', 'No name')})` }}
-      </div>
-      <div class="justify-self-end">
-        <div>
-          <CurrencyText
-            :value="accountStore.getEquityForBookId(book.id).value"
-          />
-        </div>
-        <div class="flex text-xs">
-          <CurrencyText
-            class="mr-1 text-green-500"
-            :value="
-              accountStore.getSummaryByType({
-                bookId: book.id,
-                accountType: AccountTypeType.INCOME
-              }).value
-            "
-          >
-            <template #prefix>▴</template>
-          </CurrencyText>
-
-          <CurrencyText
-            class="text-red-500"
-            :value="
-              accountStore.getSummaryByType({
-                bookId: book.id,
-                accountType: AccountTypeType.EXPENSE
-              }).value
-            "
-          >
-            <template #prefix>▾</template>
-          </CurrencyText>
-        </div>
-      </div>
-    </a>
+    ></BookListItem>
   </div>
 </template>
 
 <script setup lang="ts">
   import { onBeforeMount } from 'vue';
   import { generateUrl } from '@nextcloud/router';
-  import { translate as t } from '@nextcloud/l10n';
 
-  import CurrencyText from '../CurrencyText.vue';
+  import BookListItem from '../BookListItem.vue';
 
   import { useSettingService } from '../../services/settingService';
-
-  import { useAccountStore } from '../../stores/accountStore';
 
   import { useBookStore } from '../../stores/bookStore';
 
   import { useAccountService } from '../../services/accountService';
   import { useBookService } from '../../services/bookService';
 
-  import { AccountTypeType } from '../../stores/accountTypeStore';
-
-  const accountStore = useAccountStore();
   const bookStore = useBookStore();
 
   onBeforeMount(() => {